Extract common logic between router and gateway
