+++ pkg/server/router/router.go
@@ -17,7 +17,6 @@ package router
 import (
 	"bytes"
 	"encoding/json"
-	"fmt"
 	"io"
 	"io/ioutil"
 	"net/http"
@@ -25,8 +24,6 @@ import (
 	"sync"
 
 	"github.com/skygeario/skygear-server/pkg/server/logging"
-	"github.com/skygeario/skygear-server/pkg/server/skyerr"
-	"github.com/skygeario/skygear-server/pkg/server/skyversion"
 )
 
 var log = logging.LoggerEntry("router")
@@ -41,6 +38,7 @@ type pipeline struct {
 
 // Router to dispatch HTTP request to respective handler
 type Router struct {
+	commonRouter
 	actions struct {
 		sync.RWMutex
 		m map[string]pipeline
@@ -49,14 +47,17 @@ type Router struct {
 
 // NewRouter is factory for Router
 func NewRouter() *Router {
-	return &Router{
-		struct {
+	r := &Router{
+		actions: struct {
 			sync.RWMutex
 			m map[string]pipeline
 		}{
 			m: map[string]pipeline{},
 		},
 	}
+	r.commonRouter.payloadFunc = r.newPayload
+	r.commonRouter.matchHandlerFunc = r.matchHandler
+	return r
 }
 
 // Map to register action to handle mapping
@@ -74,80 +75,7 @@ func (r *Router) Map(action string, handler Handler, preprocessors ...Processor)
 }
 
 func (r *Router) ServeHTTP(w http.ResponseWriter, req *http.Request) {
-	var (
-		httpStatus    = http.StatusOK
-		resp          Response
-		handler       Handler
-		preprocessors []Processor
-		payload       *Payload
-	)
-
-	version := strings.TrimPrefix(skyversion.Version(), "v")
-	w.Header().Set("Server", fmt.Sprintf("Skygear Server/%s", version))
-
-	resp.writer = w
-
-	defer func() {
-		if r := recover(); r != nil {
-			resp.Err = errorFromRecoveringPanic(r)
-			log.WithField("recovered", r).Errorln("panic occurred while handling request")
-		}
-
-		if !resp.written && !resp.hijacked {
-			resp.Header().Set("Content-Type", "application/json")
-			if resp.Err != nil && httpStatus >= 200 && httpStatus <= 299 {
-				resp.writer.WriteHeader(defaultStatusCode(resp.Err))
-			} else {
-				resp.writer.WriteHeader(httpStatus)
-			}
-			if err := resp.WriteEntity(resp); err != nil {
-				panic(err)
-			}
-		}
-	}()
-
-	var err error
-	payload, err = r.newPayload(req)
-	if err != nil {
-		httpStatus = http.StatusBadRequest
-		resp.Err = skyerr.NewRequestJSONInvalidErr(err)
-		return
-	}
-
-	handler, preprocessors = r.matchHandler(req, payload)
-	if handler == nil {
-		httpStatus = http.StatusNotFound
-		resp.Err = skyerr.NewError(skyerr.UndefinedOperation, "route unmatched")
-		return
-	}
-
-	httpStatus = r.callHandler(handler, preprocessors, payload, &resp)
-}
-
-func (r *Router) callHandler(handler Handler, pp []Processor, payload *Payload, resp *Response) (httpStatus int) {
-	httpStatus = http.StatusOK
-
-	defer func() {
-		if r := recover(); r != nil {
-			log.WithField("recovered", r).Errorln("panic occurred while handling request")
-
-			resp.Err = errorFromRecoveringPanic(r)
-			httpStatus = defaultStatusCode(resp.Err)
-		}
-	}()
-
-	for _, p := range pp {
-		httpStatus = p.Preprocess(payload, resp)
-		if resp.Err != nil {
-			if httpStatus == http.StatusOK {
-				httpStatus = defaultStatusCode(resp.Err)
-			}
-			return
-		}
-	}
-
-	handler.Handle(payload, resp)
-	return httpStatus
+	r.commonRouter.ServeHTTP(w, req)
 }
 
 func (r *Router) matchHandler(req *http.Request, p *Payload) (h Handler, pp []Processor) {
